Major SQL cleanup
This patch removes a lot of unnecessary and partly even dangerous SQL
code like use statements, drop statements, unused delimiter, …

Furthermore, the preparation tables for media and series (latest and
published) are removed as well as all related triggers and are replaces
by views. This brings a lot more flexibility and clarity with almost no
loss of speed.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>